Rename : 폴더명 소문자 변경
- chatting, chatting_detail, chatting_list
- product_card
